probably better separate abi json to a single file
